Move selection operators implementations behind feature gate
